rename: common/ui로 path 이동
